move module solaris::unit to solaris::wei
