+++ solaris/src/lib.rs
@@ -13,7 +13,7 @@ mod trace;
 
 pub mod evm;
 pub mod sol;
-pub mod unit;
+pub mod wei;
 
 lazy_static! {
     pub static ref FOUNDATION: ethcore::spec::Spec = ethcore::ethereum::new_foundation(&::std::env::temp_dir());
